added abstract class component and refactored class task and task-edit
